Move MaterialSupportLayer into it own widget
