+++ lib/src/support/material_support_layer.dart
@@ -5,43 +5,6 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
-/// Wrapper with default that most widgets required that are now wrapped by a
-/// [WidgetsApp]
-class NotAWidgetsApp extends StatefulWidget {
-  const NotAWidgetsApp({
-    required this.child,
-    this.textDirection,
-    Key? key,
-  }) : super(key: key);
-
-  final Widget child;
-
-  final TextDirection? textDirection;
-
-  @override
-  State<NotAWidgetsApp> createState() => _NotAWidgetsAppState();
-}
-
-class _NotAWidgetsAppState extends State<NotAWidgetsApp> {
-  @override
-  Widget build(BuildContext context) {
-    Widget child = widget.child;
-
-    // Any Text requires a directionality
-    child = Directionality(
-      textDirection: widget.textDirection ?? TextDirection.ltr,
-      child: child,
-    );
-
-    // Inject a MediaQuery with information from the app window
-    child = MediaQuery.fromWindow(
-      child: child,
-    );
-
-    return child;
-  }
-}
-
 /// Provides parent Widgets that are required by material widgets that are
 /// not wrapped into an [MaterialApp]
 class MaterialSupportLayer extends StatefulWidget {
